% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAShiny.R
\name{TAShiny}
\alias{TAShiny}
\title{Launch TAShiny Interface}
\usage{
TAShiny()
}
\value{
Nothing
}
\description{
TAShiny() loads interactive user interface built using R shiny.
}
\details{
Start TAShiny

The interactive user interface is to provide an easy way for people who are learning Text Mining and Analytics. Includes example data for testing out a few example analysis.
}
\examples{
\dontrun{
library(shiny)
TAShiny()
}
}
\keyword{TAShiny}
